+++ NEWS.md
@@ -47,3 +47,4 @@
 - Fixed bringing main window under active secondary window when executing background command (#11407)
 - Fix for schema version comparison that breaks db in downgrade -> upgrade scenarios (rstudio-pro#3572)
 - Fixed an issue in the Electron build of the IDE on Macs where users could not clone a git repository via password-protected SSH or HTTPS (#11693)
+- Fixed scroll speed sensitivity for Mac and Linux and added a preference to adjust it (#11578)